[dev.cc] runtime: change set_sec to take int64
Fixes build.
Tested that all these systems can make.bash.

TBR=austin
CC=golang-codereviews
https://golang.org/cl/177770043

+++ src/runtime/defs_freebsd_arm.go
@@ -157,8 +157,8 @@ type timespec struct {
 	pad_cgo_0 [4]byte
 }
 
-func (ts *timespec) set_sec(x int32) {
-	ts.tv_sec = int64(x)
+func (ts *timespec) set_sec(x int64) {
+	ts.tv_sec = x
 }
 
 type timeval struct {
